+++ src/console/controllers/SectionsController.php
@@ -9,6 +9,7 @@ namespace craft\console\controllers;
 
 use Craft;
 use craft\console\Controller;
+use craft\elements\Entry;
 use craft\helpers\ArrayHelper;
 use craft\helpers\StringHelper;
 use craft\models\CategoryGroup_SiteSettings;
@@ -91,7 +92,10 @@ class SectionsController extends Controller
             return ExitCode::UNSPECIFIED_ERROR;
         }
 
-        $section = new Section();
+        $section = new Section([
+            // Avoid the default preview target
+            'previewTargets' => [],
+        ]);
 
         $validateAttribute = function($attributes, ?string &$error = null): bool {
             $section = new Section($attributes);
@@ -181,6 +185,18 @@ class SectionsController extends Controller
                 ]),
                 Craft::$app->getSites()->getAllSites(true),
             ));
+        }
+
+        $hasUrls = ArrayHelper::contains($section->getSiteSettings(), fn(Section_SiteSettings $siteSettings) => $siteSettings->hasUrls);
+        if ($hasUrls) {
+            $section->previewTargets = [
+                [
+                    'label' => Craft::t('app', 'Primary {type} page', [
+                        'type' => StringHelper::toLowerCase(Entry::displayName()),
+                    ]),
+                    'urlFormat' => '{url}',
+                ],
+            ];
         }
 
         $this->do('Saving the section', function() use ($section) {